requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
Markdown==3.4.1
mysqlclient==2.1.1
phonenumbers==8.13.6
phonenumberslite==8.13.6
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
